Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
chromadb==0.4.18
sentence-transformers==2.2.2
torch==2.1.1
transformers==4.35.2
peft==0.7.1
python-dotenv==1.0.0